pyproject.toml
[project]
name = "sync-habitica-completed-tasks-with-google-sheets"
version = "1.0.0"
description = "sync-habitica-completed-tasks-with-google-sheets"
readme = "README.md"
dependencies = [
    "google-api-core>=2.29.0",
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]


[tool.ruff]
line-length = 88

[tool.ruff.per-file-ignores]
"__init__.py" = ["F403", "E402"]
